Rename for backwards compatibility
